fix(dev-stack): wrong storage port for webhook
In previous patch we fix the storage port in traefik, but we didn't
change the URI for the webhook service which stops storage service from
saving data. This patch fix storage port in STORAGE_WEBHOOK_URI
environment variable for authenticator service. This will make new user
accounts be able to save data in storage service again.