feat(cli): add environment argument
KNoT Cloud CLI doesn't allow the user to chose wether to create a
production or development stack -- instead, it replicates a
development stack, everytime.

Update the CLI to enable the user to use 'dev' or 'prod' argument
in order to replicate the stack environment they prefer. If no
argument is passed for 'env', the CLI defaults to 'dev'.