refactor(core): Improve production stack files
Considering the 'base' definition is already suitable for all-in-one
scenarios, it's only necessary to include additional placement settings
when composing the production stack with the 'base'. Thus, this patch
updates it to remove duplicated definitions and adds the traefik config
for production environments. Also, env files were updated to remove
development only variables.